added common workspace for shared types
