+++ common/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "common",
+  "version": "1.0.0",
+  "private": true,
+  "devDependencies": {
+    "typescript": "^5.4.5"
+  }
+}
