+++ package.json
@@ -8,6 +8,7 @@
   },
   "workspaces": [
     "ebs",
-    "frontend"
+    "frontend",
+    "common"
   ]
 }
